# Fix update rate issues by working around MutliThreadedExecutor