@bmagyar 2020-12-19
Could you please elaborate on the effects of this approach?