@Rayman 2020-12-21
@bmagyar I've added a simple benchmark to the description