@Karsten1987 2021-01-13
This PR actually broke the build on OSX.
```
/Users/karsten/workspace/ros2/ros2_control_ws/src/ros-controls/ros2_control/controller_manager/src/ros2_control_node.cpp:53:9: error: use of undeclared identifier 'clock_nanosleep'
        clock_nanosleep(CLOCK_MONOTONIC, 0, &duration, NULL);
``