@Rayman 2021-01-13
Sorry, I did not know OSX was a requirment. Should there be a CI for that?

Good suggestion to replace the call with `std::this_thread::sleep_for`. It should function the same and is more portable.